+++ index.html
@@ -11,7 +11,7 @@
             name="viewport"
             content="width=device-width, initial-scale=1.0"
         />
-        <title>Vite + Vue + TS</title>
+        <title>Cryptonomicon</title>
     </head>
     <body>
         <div id="app"></div>
