+++ src/utils/constants.ts
@@ -5,8 +5,4 @@
 export const isCorrectEnum: <T>(e: T) => (token: unknown) => token is T[keyof T] =
     <T>(e: T) =>
     (token: unknown): token is T[keyof T] =>
-        Object.values(
-            e as {
-                [key: string]: unknown;
-            }
-        ).includes(token as T[keyof T]);
+        Object.values(e as { [key: string]: unknown }).includes(token as T[keyof T]);
